Resolved php warnings

+++ classes/class-uagb-block-helper.php
@@ -302,12 +302,12 @@ if ( ! class_exists( 'UAGB_Block_Helper' ) ) {
 
 			$selectors = array(
 				'.uagb-post__items'                        => array(
-					'margin-right' => UAGB_Helper::get_css_value( ( -$attr['rowGap'] / 2 ), $attr['rowGapUnit'] ),
-					'margin-left'  => UAGB_Helper::get_css_value( ( -$attr['rowGap'] / 2 ), $attr['rowGapUnit'] ),
+					'margin-right' => UAGB_Helper::get_css_value( (int)( -$attr['rowGap'] / 2 ), $attr['rowGapUnit'] ),
+					'margin-left'  => UAGB_Helper::get_css_value( (int)( -$attr['rowGap'] / 2 ), $attr['rowGapUnit'] ),
 				),
 				'.uagb-post__items article'                => array(
-					'padding-right' => UAGB_Helper::get_css_value( ( $attr['rowGap'] / 2 ), $attr['rowGapUnit'] ),
-					'padding-left'  => UAGB_Helper::get_css_value( ( $attr['rowGap'] / 2 ), $attr['rowGapUnit'] ),
+					'padding-right' => UAGB_Helper::get_css_value( (int)( $attr['rowGap'] / 2 ), $attr['rowGapUnit'] ),
+					'padding-left'  => UAGB_Helper::get_css_value( (int)( $attr['rowGap'] / 2 ), $attr['rowGapUnit'] ),
 					'margin-bottom' => UAGB_Helper::get_css_value( ( $attr['columnGap'] ), $attr['columnGapUnit'] ),
 				),
 				' .uagb-post__inner-wrap'                  => array(
@@ -341,7 +341,7 @@ if ( ! class_exists( 'UAGB_Block_Helper' ) ) {
 				),
 				' .uagb-post__image:before'                => array(
 					'background-color' => $attr['bgOverlayColor'],
-					'opacity'          => ( $attr['overlayOpacity'] / 100 ),
+					'opacity'          => ( (int)$attr['overlayOpacity'] / 100 ),
 				),
 			);
 
@@ -421,12 +421,12 @@ if ( ! class_exists( 'UAGB_Block_Helper' ) ) {
 
 			$m_selector = array(
 				'.uagb-post__items'         => array(
-					'margin-right' => UAGB_Helper::get_css_value( ( -$rowGapMobile / 2 ), $attr['rowGapUnit'] ),
-					'margin-left'  => UAGB_Helper::get_css_value( ( -$rowGapMobile / 2 ), $attr['rowGapUnit'] ),
+					'margin-right' => UAGB_Helper::get_css_value( ( -(int)$rowGapMobile / 2 ), $attr['rowGapUnit'] ),
+					'margin-left'  => UAGB_Helper::get_css_value( ( -(int)$rowGapMobile / 2 ), $attr['rowGapUnit'] ),
 				),
 				'.uagb-post__items article' => array(
-					'padding-right' => UAGB_Helper::get_css_value( ( $rowGapMobile / 2 ), $attr['rowGapUnit'] ),
-					'padding-left'  => UAGB_Helper::get_css_value( ( $rowGapMobile / 2 ), $attr['rowGapUnit'] ),
+					'padding-right' => UAGB_Helper::get_css_value( ( (int)$rowGapMobile / 2 ), $attr['rowGapUnit'] ),
+					'padding-left'  => UAGB_Helper::get_css_value( ( (int)$rowGapMobile / 2 ), $attr['rowGapUnit'] ),
 					'margin-bottom' => UAGB_Helper::get_css_value( ( $columnGapMobile ), $attr['columnGapUnit'] ),
 				),
 				' .uagb-post__inner-wrap .uagb-post__text.uagb-post__cta:last-child' => array(
@@ -485,12 +485,12 @@ if ( ! class_exists( 'UAGB_Block_Helper' ) ) {
 
 			$t_selector = array(
 				'.uagb-post__items'         => array(
-					'margin-right' => UAGB_Helper::get_css_value( ( -$rowGapTablet / 2 ), $attr['rowGapUnit'] ),
-					'margin-left'  => UAGB_Helper::get_css_value( ( -$rowGapTablet / 2 ), $attr['rowGapUnit'] ),
+					'margin-right' => UAGB_Helper::get_css_value( ( -(int)$rowGapTablet / 2 ), $attr['rowGapUnit'] ),
+					'margin-left'  => UAGB_Helper::get_css_value( ( -(int)$rowGapTablet / 2 ), $attr['rowGapUnit'] ),
 				),
 				'.uagb-post__items article' => array(
-					'padding-right' => UAGB_Helper::get_css_value( ( $rowGapTablet / 2 ), $attr['rowGapUnit'] ),
-					'padding-left'  => UAGB_Helper::get_css_value( ( $rowGapTablet / 2 ), $attr['rowGapUnit'] ),
+					'padding-right' => UAGB_Helper::get_css_value( ( (int)$rowGapTablet / 2 ), $attr['rowGapUnit'] ),
+					'padding-left'  => UAGB_Helper::get_css_value( ( (int)$rowGapTablet / 2 ), $attr['rowGapUnit'] ),
 					'margin-bottom' => UAGB_Helper::get_css_value( ( $columnGapTablet ), $attr['columnGapUnit'] ),
 				),
 				' .uagb-post__inner-wrap .uagb-post__text.uagb-post__cta:last-child' => array(
